# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Lead Engine CRE API
  description: |
    REST API for the Lead Engine CRE marketplace — a decentralized lead exchange with
    sealed-bid auctions, ZK-verified compliance, and x402 USDC settlement.

    ## Authentication
    All endpoints require a JWT Bearer token obtained via the wallet auth flow:
    1. `GET /api/v1/auth/nonce/{address}` — get signing message
    2. `POST /api/v1/auth/wallet` — submit signature, receive JWT
    3. Include `Authorization: Bearer <token>` on all subsequent requests

    ## Lead Sources
    Leads can be submitted via three channels:
    - **PLATFORM** — the built-in web form
    - **API** — programmatic submission (this endpoint)
    - **OFFSITE** — hosted landers and webhook integrations
  version: 1.0.0
  contact:
    name: Lead Engine CRE
    url: https://github.com/bnmbnmai/lead-engine-cre
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://lead-engine-cre-api.onrender.com
    description: Production
...
paths:
...
  /api/v1/leads:
    get:
      tags: [Leads]
      summary: List leads
      security:
        - bearerAuth: []
      parameters:
        - name: vertical
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_AUCTION, IN_AUCTION, REVEAL_PHASE, SOLD, EXPIRED]
        - name: state
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, reservePrice, auctionEndAt]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated list of leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/Pagination"
...
  /api/v1/bids:
    post:
      tags: [Bids]
      summary: Place a bid
      description: |
        Submit a sealed bid on a lead. Supports two modes:
        - **Sealed bid**: provide `commitment` (keccak256 hash of amount + salt)
        - **Direct bid**: provide `amount` directly (for non-ZK auctions)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/BidCommit"
                - $ref: "#/components/schemas/BidDirect"
      responses:
        "201":
          description: Bid placed
        "400":
          description: Invalid bid
...
  /api/v1/analytics/overview:
    get:
      tags: [Analytics]
      summary: Dashboard overview stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Overview statistics
  /api/v1/analytics/leads:
    get:
      tags: [Analytics]
      summary: Lead analytics
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: vertical
          in: query
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
      responses:
        "200":
          description: Lead analytics data
  /api/v1/analytics/bids:
    get:
      tags: [Analytics]
      summary: Bid analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bid analytics data
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BidCommit:
      type: object
      required: [leadId, commitment]
      properties:
        leadId:
          type: string
        commitment:
          type: string
          pattern: "^0x[a-fA-F0-9]{64}$"
          description: keccak256(abi.encodePacked(amount, salt))
    BidDirect:
      type: object
      required: [leadId, amount]
      properties:
        leadId:
          type: string
        amount:
          type: number
          minimum: 0.01
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
...